Fix for namespace in API endpoint mismatch
Some APIs include a type specifier in the URL, like `microsoft.graph.sitePage`. For some of these, the OpenAPI that gets generated shortens this to `graph.sitePage`. This was causing a match to not be found. Added logic to try the shortened version if no match is found with the original.